lambda/requirements.txt
boto3
discord-webhook
python-dateutil
